+++ CHANGES.md
@@ -7,6 +7,7 @@
 - FIX: Don't except ImportErrors in `__init__` and therefore look for version internally in `__meta__` file
 - FIX: Fix band list type when writing a stack on disk
 - FIX: Fix nodata management when stacking a mosaic
+- FIX: Better condensed name for pairs
 
 ## 0.4.1 (2025-01-07)